--- a/editor/src/messages/portfolio/document/node_graph/document_node_definitions.rs
+++ b/editor/src/messages/portfolio/document/node_graph/document_node_definitions.rs
@@ -1613,7 +1613,7 @@ fn static_nodes() -> Vec<DocumentNodeDefinition> {
 								..Default::default()
 							}),
 						),
-						PropertiesRow::with_override(
+						InputMetadata::with_name_description_override(
 							"Tilt",
 							"Faux italic",
 							WidgetOverride::Number(NumberInputSettings {