--- a/editor/src/messages/portfolio/document/utility_types/mod.rs
+++ b/editor/src/messages/portfolio/document/utility_types/mod.rs
@@ -5,3 +5,4 @@ pub mod misc;
 pub mod network_interface;
 pub mod nodes;
 pub mod transformation;
+pub mod wires;