… editor/src/messages/portfolio/document/utility_types/network_interface.rs
… editor/src/messages/portfolio/document/utility_types/nodes.rs
… editor/src/messages/portfolio/document/utility_types/wires.rs
… editor/src/messages/portfolio/document_migration.rs
… editor/src/messages/preferences/preferences_message.rs
… editor/src/messages/preferences/preferences_message_handler.rs
… editor/src/test_utils.rs
… frontend/src/components/views/Graph.svelte
… frontend/src/messages.ts
… frontend/src/state-providers/node-graph.ts
… frontend/vite.config.ts.timestamp-1751495168688-0f7ee9d1b378f.mjs
… frontend/wasm/src/editor_api.rs
… node-graph/gcore/src/vector/click_target.rs
… node-graph/graph-craft/src/document.rs