Closes #2830.

Move wire rendering into backend
Implement culling for offscreen nodes
Improves properties/document upgrade stability
Validate that the input metadata matches the node inputs, and get populated from the definition if non existent
Consolidate logic in the widget parameters `new` function
Improve logic for collecting/sending input metadata for the frontend nodes

Decreases frame time from 120ms -> 20ms when zoomed in on red dress node graph
Increases frame time from 120 -> 200ms when fully zoomed out

Old:
![chrome_D18AToYUWg](https://github.com/user-attachments/assets/fde1d878-2be2-40e6-8c85-2cb4705f345f)
New:
![chrome_AMXWLyTqfv](https://github.com/user-attachments/assets/a7defe74-8af5-4281-9a22-baa7da92b48b)

Known issues: 
Subpath rectangle_intersections_exist function is buggy, so dragging onto wire is still inconsistent