@Keavon 2025-07-05
Found one last bug: wires break when deleting a node in the middle of a chain, undoing the deletion, then redoing the deletion.